Nav changed to link to '/search/path' for buttons. Search works.
Nav changed to link to '/search/path' for buttons. Search works.

NavLink replacements made like this: <NavLink
to='/search/batman'>Batman</NavLink>

SearchResults.js deleted.

NotFound route rendering is still an issue before first search has been
performed.

Still need to clean up application and potentially allow for a
generalized path variable so that ‘/search/‘ does not need to be
repeated in the code so many times and in case the name changes.