Search works. Debugging NotFound Route display in 2 instances.
Search works. Debugging NotFound Route display in 2 instances.

1) NotFound displays each time SearchForm is rendered without a search
result, including the first time the component is rendered, which is
incorrect as the user has yet to submit a search term.

2) If no search results are found, the title renders and uses the
search term. This search term that returns no data, for example
“aosidoiasdoiasdioasjd”, is rendered as “Aosidoiasdoiasdioasjd Images”
atop the NotFound route rendering, as it should, when no image data is
returned.